Moved several classes into Core namespace.
