+++ src/Database/ResultsCache.php
@@ -13,10 +13,10 @@
 namespace Bead\Database;
 
 use ArrayAccess;
+use Bead\Core\Application;
+use Bead\Facades\Log;
 use Countable;
 use DirectoryIterator;
-use Bead\Application;
-use Bead\Facades\Log;
 use Exception;
 use Iterator;
 use JsonException;
