+++ test/Core/files/translations/fr.csv
