+++ test/smoke/public/index.php
@@ -5,8 +5,8 @@
  * @version 0.9.2
  */
 
+use Bead\Core\WebApplication;
 use Bead\View;
-use Bead\WebApplication;
 
 require_once __DIR__ . "/../../../vendor/autoload.php";
